Refactor original matcher
